[FEATURE] Added method to run recursive requests
requestRecursive() was added to make recursive requests to the
data api. You can either add the body of each request to a
local variable that would be returned at the completion, however
this isn't recommended.

Instead you can pass a callback to requestRecursive() that is
triggered on completion of each run of the Data API.