[BUG] Prevent a 100-Continue response from the server
By default, libcurl is sending `Expect: 100-Continue` header when the
request body is above a certain size. The server MAY send a 100-Continue
header in the same stream as the final response to comply with this -
depending on the state of the request body upload.

When callers check the response object, they may see 100 as the response
status code instead of 200 in this case. This is confusing. This change
prevents it.