Offers an opt-in way to log requests make against the Data API using the SDK.
Tracks the following types:
* request
* response
* summary
* error

And logs to the following handlers:
* disk
* syslog
* stdout

You should be able to opt-in to any handler/type combination by passing configuration to the DataApi constructor.

Eg the following will log `request`, `response`, `summary` and `error` types to disk (at the `path` location), and `summary` to stdout. Nothing is logged to syslog:

```
$logOptions = [
    'handlers' => [
        'file' => [
            'path'  => __DIR__,
            'types' => [
                'request',
                'response',
                'summary',
                'error'
            ]
        ],
        'stdout' => [
            'types' => [
                'summary'
            ]
        ],
        'syslog' => []
    ]
];

$dataapi = new DataApi([], $logOptions);
```