[REFACTOR] Event init with the same form of user hash
The Init object previously expected the request to be an object with an
array `users` property,

   { "users": [ "brianmoser", "hankshrade", "jessiepinkman", "walterwhite" ] }

This structure doesn't match the Events API initialisation, which
requires the `users` property to be an associative array of user IDs to
there signature,

   { "users": { "brianmoser": "...", "hankshrader": "...", "jessiepinkman": "...", "walterwhite": "..." } }

With this change, we now expect such an associative array, and replace
whatever value was provided with the signature for each user ID provided
in the keys of that array. The easiest way to use this is simply to
provide an empty string,

   { "users": { "brianmoser": "", "hankshrader": "", "jessiepinkman": "", "walterwhite": "" } }

but anything else will work, too

   { "users": { "brianmoser": "SIGNATURE HERE", "hankshrader": "SIGNATURE HERE", "jessiepinkman": "SIGNATURE HERE", "walterwhite": "SIGNATURE HERE" } }

We nonetheless include backward compatibility code to support arrays for
backward compatibility, with a PHP warning.

Signed-off-by: Olivier Mehani <olivier.mehani@learnosity.com>