Updating the Makefile to have the following changes:
- Remove the version generation

There are 3 ways the SDK can be imported, direct github clone, require
from composer or download the pre-packaged zip. The scenario where we
enable rebuild the zip file from the pre-packaged zip is one that may
lead to more questions than answers. This change remove the ability to
rebuild as the zip file can be re-downloaded instead.

- Update the clean process.

Previously, clean does not remove old dist files as it only act on the
dist target. Since we may have old dists left over from previous builds,
the next iteration of packaging will not remove them. The change here
is to remove all the dist files and directories.