[BUILD] Fix dist-test
Previously, a race condition between dist-zip, dependent on `clean`, and
dist-test, dependent on `install-vendor-dev` led to missing vendors as
make (rightly) ran the latter before the former, leading to it then
being absent when the tests needed to run.

We now use the local Makefile to run install the dependencies and run
the tests directly into the `dist` directory (after the archive has been
built).

[BUILD] Remove deprecated composer option

Signed-off-by: Olivier Mehani <olivier.mehani@learnosity.com>